requirements.txt
tensorflow-gpu>=2.0.0
absl
numpy>=1.17.3
